[SHELL32][SETUPAPI][SCM][WIN32SS] User-mode + display-path fixes, ARM64 printing
- shell32: SHGetKnownFolderPath + SHCreateItemFromIDList (Vista+ surface); shlwapi/shcore target 0x600; propbag using-decl cleanup
- setupapi: DIRID 13 (DriverStore) resolves to system32\drivers
- scm: service-start pipe wait also watches process exit (no PipeTimeout serialization for exiting images); kill never-connected single-service images; eventlog-unavailable logged once
- services: RpcServerListen tolerates RPC_S_ALREADY_LISTENING (5 services)
- msgina: Userinit is a comma-separated command list (Windows semantics)
- explorer: AlwaysOnTop taskbar not forced HWND_BOTTOM by fullscreen rude apps
- opengl32: viewport tracks WM_SIZE growth (stale-viewport black region); ICD SetPixelFormat also records the format with win32k; per-process renderer proof line
- win32ss/eng: devmode list preserves InstalledDisplayDrivers order; EngpGetPDEV NULL-gpmdev guard; large DDBs use kernel sections (paged-pool starvation)
- printing: Windows ARM64 print environment + prtprocs/arm64 packaging
- user32: AdjustWindowRectExForDpi/EnableNonClientDpiScaling/GetSystemMetricsForDpi
- freeldr: SMBIOS 3.0 EPS handed to the kernel via the loader extension; jscript wine-sync hex_to_int; qmgr tidy; cmd VALUE type fix

+++ base/shell/cmd/cmd.c
@@ -1186,7 +1186,7 @@ GetEnhancedVar(
         {
             static const struct {
                 TCHAR Character;
-                DWORD Value;
+                WORD  Value;
             } *Attrib, Table[] = {
                 { _T('d'), FILE_ATTRIBUTE_DIRECTORY },
                 { _T('r'), FILE_ATTRIBUTE_READONLY },
